clean up python script files
